dri2,present: move in pixmap before getting addresses
First of all, we were not forcing the pixmap to be moved in before
attempting to present it. Secondly, in certain configurations, we avoid
ever backing pixmaps with vram (in GPUs with 32MB of VRAM or less).

This fixes the observed cases where we ended up assuming that a bo was
backing a pixmap without doing the explicit move-in.

Signed-off-by: Ilia Mirkin <imirkin@alum.mit.edu>